Extract function to file
